Alternative impl that finds default DependencyContext via reflection
Given that
- Assembly.GetEntryAssembly is not available until .NET standard 1.5
- DependencyContext.Default is not available until .NET standard 1.6

The only viable way to reduce the dependency on the .NET standard to
1.3 without losing functionality is to use reflection to dig this
off the concrete types (which are supplied by either .NET framework
or .NET core). As both .NET framework and core have had these types
in since very early days, this is generally a very safe thing to do,
albeit, slightly hacky.

The final point of sadness is that the test that was introduced only
works on one platform, due to an unfixed bug in vstest [1].
Still, the test at least proves this approach works on .NET core.

Fixes #60

[1] https://github.com/Microsoft/vstest/issues/649